.github/actions/validate/requirements.txt
pyyaml
owasp_schema